Add shared OAuth provider registry and token persistence module
